# docs/swagger.yaml
basePath: /
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  tables.Tag:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      tag_created:
        type: string
      tag_data:
        type: string
      tag_exclude:
        type: string
      tag_id:
        type: string
      tag_name:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: collector-api-contact@opensvc.com
    name: OpenSVC SAS
    url: https://www.opensvc.com
  description: Organization clusters, nodes, services and more.
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: OpenSVC collector API
  version: "1.0"
paths:
...
  /tags/{id}:
    delete:
      consumes:
      - application/json
      description: |-
        Delete a tag by index, uuid or name.
        Requires the TagManager privilege.
        Cascade deletes the tag attachements to nodes and services.
      parameters:
      - description: the index of the entry in database, or uuid, or name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tables.Tag'
            type: array
        "204":
          description: No Content
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Delete a tag
      tags:
      - tags
    get:
      consumes:
      - application/json
      parameters:
      - description: the index of the entry in database, or uuid, or name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tables.Tag'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Show a tag
      tags:
      - tags
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"